multi: remove RefCell from LndNodeSigner
Initially a RefCell was added to the LndNodeSigner because the original version
of tonic_lnd returns a mutable reference to the signer. However this reference
doesn't need to be mutable:
https://github.com/orbitalturtle/tonic_lnd/pull/4

For this PR, this was needed so we could spin up the onion messenger and server
in a separate thread in the integration tests. But it's a change that could
help with future efficiency as well because the onion messenger can now be used
in a multi-threaded way.