server: default lndk conf in lndk dir
LNDK is already using a datadir to store TLS certificates.
In this PR we default LNDK config and data to ~/.lndk.
Datadir is moved inside the default dir ~/.lndk/data. Then,
we expect lndk.conf file to be iether in root of the project,
or the default LNDK config folder or --config flag.
Precedence is flag > root > default dir.