handler: extract `send_payment` and `track_payment` to `lnd_requests`
This commit moves `send_payment` and creates a `track_payment`
function in `lnd_request` module.

`send_payment` has the responsibillity to interact with lnd to
 query for possible routes and then send the payment to the one
found. Key difference with previous implementation is that it does
not interact with active payment state of the handler.

`track_payment` is added to encapsulates the lnd call of tracking
the payment itself.