tests: test get_invoice flow on integration tests
This commit changes `test_lndk_send_invoice_request` to
`test_lndk_get_invoice`. Main difference is that we now wait to
receive the invoice back when sneding an invoice request.
Also, adds a helper function `get_invoice_with_retry` to retry
in case of failure because there is a race condition between
the internal graph and when wecall get_invoice.