messenger: add a integration test to check reconnection.
This commit adds a integration test for checking that when a LND node is
not available the reconnection wrapper will keep trying to reconnect
making possible to use the same handler for trying to pay an offer.
When the LND is broken the first pay offer will fail whereas when the
reconnection succeed the second attempt of pay a offer will finish.
It was also necessary to abstract how LND arguments are generated in order
to restart LND node using previous arguments.