tests: add `wait_for_addresses_to_sync` helper
To test out when peers are not connected we need the graph to sync
with the node addresses to then be connected. The method added
waits for 100s for an specific node to appear in the LND graph.

Finally, as we need more freedom when using `connect_network` we
add a boolean flag ti announce or not the ldk1 <-> ldk2 connection.